tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/size.attributes.removed.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: size.attributes.removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.removed</h1>
<p class="desc">Removing content attributes reverts to default size</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.removed.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Removing content attributes reverts to default size");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 120, "canvas.width", "120");
canvas.removeAttribute('width');
_assertSame(canvas.width, 300, "canvas.width", "300");


});
</script>
